Fix Neo4jContainer
The container didn't expose the ports required to be started up, and would therefor never start. We've added fixed ports. In the future it may be possible to add dynamic ports so that we can start multiple containers at the same time, but this would presently require quite a lot of work.